begin implement verb `export`
let's the user open the GUI, select all items there are to select, then submits the GUIs and writes the configuration to file.

After this, we must implement a verb 'import' which loads the config properly. This verb shall by default load the config, then execute it.
Providing the flag `--edit` to it will let the user load and modify the configuration. In this case, this mode will behave as `export` after all GUIs are submitted.